algo/tests: split tests into unit and integration
